requirements.txt
Flask
requests
pytrends
praw
google-api-python-client
python-dotenv
apscheduler
sqlalchemy

